worker/pyproject.toml
[project]
name = "quiz-worker"
version = "0.1.0"
description = "Generates quizzes and saves them to Supabase"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "langchain>=1.1.2",
    "langchain-openai>=1.1.0",
    "pyjwt>=2.10.1",
    "uvicorn>=0.38.0",
]
